+++ authlete/dto/userinfo_response.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -32,7 +32,8 @@ class UserInfoResponse(ApiResponse):
             'responseContent':       str,
             'properties':            Property,  # list of Property
             'clientIdAlias':         str,
-            'clientIdAliasUsed':     bool
+            'clientIdAliasUsed':     bool,
+            'userInfoClaims':        str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)